REFACTOR files and code
